requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib==1.7.4
celery==5.3.5
redis==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
alembic==1.12.1
supabase==1.0.4
bcrypt==4.0.1
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
aiosmtplib==2.0.2
jinja2==3.1.2
pandas==2.1.3
plotly==5.18.0
